Adapt code to new changes.
